Phase B-4: kx-context-assembler lib.rs split (Rule 3, behavior-preserving)
Fourth Phase B refactor per Rule 3. Pure move; zero behavior change.

Before: src/lib.rs 949 lines (borderline by line count; small public
surface but the inline test module is ~540 lines).

After: thin src/lib.rs 105 lines (crate attrs + crate doc + mod decls +
pub use re-exports + #[cfg(test)] mod tests;) + 4 sibling modules:

- types.rs (AssembledItem + AssembledContext + impls)
- errors.rs (AssemblyError)
- assemble.rs (the pure assemble fn)
- tests.rs (the 11 inline unit tests)

Re-exports preserve every previously-public path (kx_context_assembler::{
AssembledItem, AssembledContext, AssemblyError, assemble}).

Tests.rs imports tightened during the move: the original `use super::*;`
brought in the lib.rs's `use kx_content::{ContentRef, ContentStore}`,
`use kx_mote::{EdgeKind, Mote, MoteId}`, `use bytes::Bytes`, etc.
transitively. After extraction, only ContentRef + ContentStore + Bytes +
Mote + MoteId actually appeared in test bodies — added those explicitly.
EdgeKind, Snapshot, ResolutionError were transitively-only — omitted.

Local gates green on Apple Silicon (--all-features):
- cargo build -p kx-context-assembler: clean
- cargo test --workspace --all-features: 645 + 8 (IDENTICAL)
- cargo test -p kx-context-assembler: 11 + 3 + 12 + 5 = 31 (IDENTICAL)
- cargo clippy -p kx-context-assembler --all-targets -- -D warnings: clean
- cargo fmt --all -- --check: clean

Next Phase B slice: kx-capability (1304 lines; D24 broker seam = max-strict
per Rule 1).
